+++ src/users/domain/UserRepository.go
@@ -1,5 +1,5 @@
 package domain
 
 type UserRepository interface {
-	GetUser(id int64) User
+	GetUser(id int) User
 }
